+++ NEWS.md
@@ -4,6 +4,8 @@
 - `NodeNumbers()` returns the indices of nodes within a tree.
 - `AddTip(edgeLength = NULL)` defaults to `lengthBelow`. This will become the
   default in a future release.
+- An entry point to the C++ function `root_on_node()` is now exported
+  (intended for expert use only).
 
 # TreeTools 1.10.0 (2023-08-18) #