+++ NEWS.md
@@ -1,5 +1,6 @@
-# TreeTools 1.10.0.9001 (development) #
+# TreeTools 1.10.0.9002 (development) #
 
+- `YuleTree()` generates a random tree by the Yule process.
 - `DescendantTips()` complements `DescendantEdges()`.
 - `NodeNumbers()` returns the indices of nodes within a tree.
 - `AddTip(edgeLength = NULL)` defaults to `lengthBelow`. This will become the